refactor(layout): move validate + plan_diff into vmux_layout::reconcile
